+++ aiohttp/connector.py
@@ -104,14 +104,19 @@ class BaseConnector(object):
     force_close - Set to True to force close and do reconnect
         after each request (and between redirects).
     limit - The limit of simultaneous connections to the same endpoint.
+    disable_cleanup_closed - Disable clean-up closed ssl transports.
     loop - Optional event loop.
     """
 
     _closed = True  # prevent AttributeError in __del__ if ctor was failed
     _source_traceback = None
 
+    # abort transport after 2 seconds (cleanup broken connections)
+    _cleanup_closed_period = 2.0
+
     def __init__(self, *, conn_timeout=None, keepalive_timeout=sentinel,
-                 force_close=False, limit=20, time_service=None, loop=None):
+                 force_close=False, limit=20, time_service=None,
+                 disable_cleanup_closed=False, loop=None):
 
         if force_close:
             if keepalive_timeout is not None and \
@@ -151,10 +156,17 @@ class BaseConnector(object):
 
         self.cookies = SimpleCookie()
 
+        # start keep-alive connection cleanup task
         self._cleanup_handle = None
         if (keepalive_timeout is not sentinel and
                 keepalive_timeout is not None):
             self._cleanup()
+
+        # start cleanup closed transports task
+        self._cleanup_closed_handle = None
+        self._cleanup_closed_disabled = disable_cleanup_closed
+        self._cleanup_closed_transports = []
+        self._cleanup_closed()
 
     def __del__(self, _warnings=warnings):
         if self._closed:
@@ -219,6 +231,10 @@ class BaseConnector(object):
                         if proto.is_connected():
                             if use_time - deadline < 0:
                                 transport.close()
+                                if (key[-1] and
+                                        not self._cleanup_closed_disabled):
+                                    self._cleanup_closed_transports.append(
+                                        transport)
                             else:
                                 alive.append((transport, proto, use_time))
 
@@ -229,6 +245,22 @@ class BaseConnector(object):
 
         self._cleanup_handle = self._time_service.call_later(
             self._keepalive_timeout / 2.0, self._cleanup)
+
+    def _cleanup_closed(self):
+        """Double confirmation for transport close.
+        Some broken ssl servers may leave socket open without proper close.
+        """
+        if self._cleanup_closed_handle:
+            self._cleanup_closed_handle.close()
+
+        for transport in self._cleanup_closed_transports:
+            transport.abort()
+
+        self._cleanup_closed_transports = []
+
+        if not self._cleanup_closed_disabled:
+            self._cleanup_closed_handle = self._time_service.call_later(
+                self._cleanup_closed_period, self._cleanup_closed)
 
     def close(self):
         """Close all opened transports."""
@@ -252,13 +284,24 @@ class BaseConnector(object):
             for transport in chain(*self._acquired.values()):
                 transport.close()
 
+            # cacnel cleanup task
             if self._cleanup_handle:
                 self._cleanup_handle.cancel()
+
+            # cacnel cleanup close task
+            if self._cleanup_closed_handle:
+                self._cleanup_closed_handle.cancel()
+
+            for transport in self._cleanup_closed_transports:
+                transport.abort()
 
         finally:
             self._conns.clear()
             self._acquired.clear()
             self._cleanup_handle = None
+            self._cleanup_closed_transports.clear()
+            self._cleanup_closed_handle = None
+
         return ret
 
     @property
@@ -335,7 +378,8 @@ class BaseConnector(object):
             if transport is not None and proto.is_connected():
                 if t1 - t0 > self._keepalive_timeout:
                     transport.close()
-                    transport = None
+                    if key[-1] and not self._cleanup_closed_disabled:
+                        self._cleanup_closed_transports.append(transport)
                 else:
                     if not conns:
                         # The very last connection was reclaimed: drop the key
@@ -391,6 +435,9 @@ class BaseConnector(object):
         reader = protocol.reader
         if should_close or (reader.output and not reader.output.at_eof()):
             transport.close()
+
+            if key[-1] and not self._cleanup_closed_disabled:
+                self._cleanup_closed_transports.append(transport)
         else:
             conns = self._conns.get(key)
             if conns is None: